+++ doc/changelog.rst
@@ -43,7 +43,7 @@ See the `PyMongo 4.15.4 release notes in JIRA`_ for the list of resolved issues
 in this release.
 
 .. _PyMongo 4.15.4 release notes in JIRA: https://jira.mongodb.org/secure/ReleaseNote.jspa?projectId=10004&version=47237
->>>>>>> master
+
 
 Changes in Version 4.15.3 (2025/10/07)
 --------------------------------------